Refactor: extract subscribers
