+++ simple-publish-subscribe/src/main/java/com/ksamorodov/task/saber/Application.java
@@ -3,6 +3,8 @@ package com.ksamorodov.task.saber;
 import com.ksamorodov.task.saber.publisher.Publisher;
 import com.ksamorodov.task.saber.strategy.PublishingStrategy;
 import com.ksamorodov.task.saber.strategy.impl.BroadcastStrategy;
+import com.ksamorodov.task.saber.subscribers.Subscriber1;
+import com.ksamorodov.task.saber.subscribers.Subscriber2;
 
 import java.util.concurrent.Flow;
 
@@ -13,51 +15,9 @@ public class Application {
         Publisher<String> publisher = new Publisher<>(broadcastStrategy);
 
         // создаем двух подписчиков
-        Flow.Subscriber<String> subscriber1 = new Flow.Subscriber<String>() {
-            @Override
-            public void onSubscribe(Flow.Subscription subscription) {
-                System.out.println("Subscriber 2 is subscribed.");
-                subscription.request(1);
-            }
+        Flow.Subscriber<String> subscriber1 = new Subscriber1();
 
-            @Override
-            public void onNext(String item) {
-                System.out.println("Subscriber 2 received item: " + item);
-            }
-
-            @Override
-            public void onError(Throwable throwable) {
-                System.out.println("Subscriber 2 received an error.");
-            }
-
-            @Override
-            public void onComplete() {
-                System.out.println("Subscriber 2 is complete.");
-            }
-        };
-
-        Flow.Subscriber<String> subscriber2 = new Flow.Subscriber<String>() {
-            @Override
-            public void onSubscribe(Flow.Subscription subscription) {
-                System.out.println("Subscriber 2 is subscribed.");
-                subscription.request(1);
-            }
-
-            @Override
-            public void onNext(String item) {
-                System.out.println("Subscriber 2 received item: " + item);
-            }
-
-            @Override
-            public void onError(Throwable throwable) {
-                System.out.println("Subscriber 2 received an error.");
-            }
-
-            @Override
-            public void onComplete() {
-                System.out.println("Subscriber 2 is complete.");
-            }
-        };
+        Flow.Subscriber<String> subscriber2 = new Subscriber2();
 
         // подписываем подписчиков на издателя
         publisher.subscribe(subscriber1);
